Fix Solr Indexing on Questions and Responses and Also Don't Include Spam Questions on Search
Disabled the solr hook for reindexing questions on save, and put in a
custom hook to reindex questions on create and added another custom
hook to reindex the parent question (which reindexes it's responses)
when a question has a response so that the question and the associated
responses stay up to date in the solr index. This is to resolve an
issue with solr not indexing a question's responses on save b/c it's an
association and the order of operations of save was preventing that.