Image Upload for Expert Responses
Thought I would work on this before the final release instead of after
and see if we can include this on day one. Needs a little style work,
and doesn't have any image deletion capabilities with it or ability to
add images to responses after the response is submitted, which might be
a debatable feature.