More Search Changes
Included in this commit and the previous search commit are changes that
make a paginated global public question search, an expanded search on
user bios and group descriptions, and a modification to the search on
experts ordering them by the last active date. An index rebuild is
required for this and some style work is needed on the public search
area.