Fixed Submitter Followup Issue and Issue With Multiple Answerers
The followup issue is that if a group was assigned a question and
someone answered it, then a person from the group followed up, the last
assignee (who was the submitter) was not being detected to it could be
assigned back to that person. The issue with multiple answerers was
that if someone tried to respond to a question that had already been
responded to while they were answering the question, it was not being
handled correctly. All this has been fixed along with fixing some of
the flash error references within the application.