Fixed Duplicate Group Entry App Error
Occurred when an expert tried to create a group with the same name as
an existing group with different case for the name. The rails
validation was case sensitive, but mysql is not when validating unique
keys thus causing the rails validation to pass but the mysql constraint
to fail.