Remove Group Filter and Tag Filter from Listviews if Group and Tag are Removed
This handles the case that someone selects from their groups a group to
filter listviews by and/or selects a tag from their tags to filter by
and then leaves the group or removes the tag. Now upon removal of said
group or tag, the listivew filter is adjusted accordingly. Also
included are some changes to how the filter choices are presented on
the listview pages so that now they show up more clearly.