Fixes to Reports Queries and Addition of New Column for Determining If a Question Has Ever Been Resolved
Also included are other tweaks and guards for reports. In addition,
added the update of the ever_answered column when a question is
resolved to keep that up to date. Turns out this new column will not
help in the reports queries that much currently b/c we still have to
join with question_events to get the time the question was resolved in
the timeframe specified, but this column will likely help with other
reporting efforts we have moving forward.