Clearing the working_on_this Timestamp on Question Resolve
Adding this as an extra edge case guard in case someone works on a
question, resolves it, then reactivates it by having it assigned to
oneself. In that case, the assign_to code does not kick off the code to
clear the timestamp. This should help fill any potential holes that
could keep old timestamps around.