Change How Assignments from Submitter Responses are Done Based on Away Status
If a question is submitted, it gets assigned, the assignee goes on away
status, the submitter posts another response, then the question gets
automatically reassigned to another expert based on group leadership or
a question wrangler. The same thing happens if an expert resolves a
question, goes on away status, then the submitter follows back up with
a response, it will be automatically reassigned to another available
expert based on group leadership or a question wrangler.