Added Option for Editor of Response to Notify Question Submitter and Added Additional Notifications
Also added notification to author of response on every update along
with notification logic and added extra column in revision history that
lets the viewer know if a response edit notification was sent to the
submitter of the question or not. Still a work in progress. NOTE: since
this has not been deployed anywhere yet, one of the migrations was
edited. If you migrated the previous version locally, you'll need to
update your schema and re-migrate.