Added Notifications to the Question Submitter on Response Edit
This only occurs when the expert elects to send a response to the
question submitter. Also did some tweaking on how the submitter
notification data is saved and displayed. NOTE: Along with this,
tweaked the existing migration since it's not deployed anywhere, but
you will need to make adjustments if you had migrated locally.